runpod/requirements.txt
# RunPod extraction worker — dependencias
# Python 3.10 + CUDA 12.1
# Espejo del backend local MENOS lo que no se usa en el pod (gmail, UI, auth).

# FastAPI server
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# SQLAlchemy (para DB SQLite in-memory)
sqlalchemy>=2.0.23
pydantic-settings>=2.0

# Document extraction tier 1-3
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
pdf2image>=1.16
pytesseract>=0.3.10
Pillow>=10.0
pdftext>=0.3.0

# OCR con GPU — instalados por separado en setup.sh con el index oficial
# de PaddlePaddle (wheels públicos en PyPI no tienen builds para CUDA 12.x).
# paddlepaddle-gpu  ← instalado en setup.sh
# paddleocr         ← instalado en setup.sh

# Layout/tablas (GPU)
marker-pdf>=1.10.0

# NER
spacy>=3.7.0

# AI providers (para ai_extractor si lo llama el pipeline)
anthropic>=0.40.0
openai>=1.50.0

# PII redaction (Presidio opcional, usamos spaCy directo)
cryptography>=41.0

# Utilidades
python-dateutil>=2.8.2
httpx>=0.25.0
